Remove tab chars, handle GetTempPath failure on Windows.

+++ arbiter/util/md5.cpp
@@ -158,14 +158,14 @@ void md5_final(Md5Context *ctx, uint8_t hash[])
 
     // Append to the padding the total message's length in bits and transform.
     ctx->bitlen += ctx->datalen * 8;
-    ctx->data[56] = ctx->bitlen;
-    ctx->data[57] = ctx->bitlen >> 8;
-    ctx->data[58] = ctx->bitlen >> 16;
-    ctx->data[59] = ctx->bitlen >> 24;
-    ctx->data[60] = ctx->bitlen >> 32;
-    ctx->data[61] = ctx->bitlen >> 40;
-    ctx->data[62] = ctx->bitlen >> 48;
-    ctx->data[63] = ctx->bitlen >> 56;
+    ctx->data[56] = static_cast<uint8_t>(ctx->bitlen);
+    ctx->data[57] = static_cast<uint8_t>(ctx->bitlen >> 8);
+    ctx->data[58] = static_cast<uint8_t>(ctx->bitlen >> 16);
+    ctx->data[59] = static_cast<uint8_t>(ctx->bitlen >> 24);
+    ctx->data[60] = static_cast<uint8_t>(ctx->bitlen >> 32);
+    ctx->data[61] = static_cast<uint8_t>(ctx->bitlen >> 40);
+    ctx->data[62] = static_cast<uint8_t>(ctx->bitlen >> 48);
+    ctx->data[63] = static_cast<uint8_t>(ctx->bitlen >> 56);
     md5_transform(ctx, ctx->data);
 
     // Since this implementation uses little endian byte ordering and MD uses
